uploader.py won't exit if some of the libraries fail
The root cause with this is that updating the documentation of 'Remote' test
library actually tries to open a connection to the remote server to get
the keywords of the remote library. This indeed fails if the remote server
is not running.

Anyway, the uploader.py was modified to work so that if updating some library
fails, the script won't exit but instead prints an error and continues to the
subsequent libraries given as arguments.

Update issue 17
Status: Done