Added basic functionality to upload multiple versions of a library.
Update issue 19

Libraries are now considered to be the same if they have the same name and same version identifier. If version differs from existing library with the same name, a new version is created. All version numbers along with links to them are shown both in the listing of libraries and in the detailed library view.