upload.py: made uploader not fail on the first remote error.
Update issue 30
Status: Done

upload.py now collects all the failed file uploads in a list and prints it out at the end of execution. Uploading will not stop to the first error, instead all found/given files are tried to be uploaded even when some of the uploads fail.